fix: lint warning fix
