+++ src/pages/lost-article/edit/hook/useLostEdit.tsx
@@ -56,7 +56,6 @@ export function useLostEdit() {
       });
 
       navigate(`/lost-article?category=state`);
-      window.location.reload();
     } catch (e) {
       console.error(e);
     }
